Separate file for apt backup backend
- Create separate class for apt backup backend
  functions, and put them in a separate file
